// src/core/user/enum/user-role.enum.ts
export enum USER_ROLE {
  BLOCKED = 0,
  USER = 1,
  ADMIN = 2, // Удалять коменты и изменять их
  SUPER = 3, // Получает уведомления на почту
}

// Админ
// не получает письма на почту
// Не видит статистику
// Не видит заказы
// Не видит список пользователей